🌱 Split AlertDrillDown.tsx into orchestrator, parts, and hook
Extracts AlertDrillDown.tsx (534 lines, 17 hook calls) following the
established X.tsx + X.parts.tsx + useX.ts drilldown-split pattern.

- useAlertDrillDown.ts — all state, effects, memoized context, and
  the fetchSourceRule/handleCopy/handleDiagnose callbacks
- AlertDrillDown.parts.tsx — pure presentational sub-components
  (header, tab bar, overview/labels/source/ai tab panels) plus the
  getSeverityStyle/getStateStyle style helpers
- AlertDrillDown.tsx — thin orchestrator (123 lines, 2 hook calls)

No behaviour change — pure extraction. Drill-down registration and
useDrillDown() props unchanged.
